feat(backend): harden infrastructure with governance, security, and sub-account features
Implementation summary:
- Governance: Integrated global Config-driven safety checks (maintenance mode, withdrawal toggles) across WithdrawalProcessor and service layers.
- Security: Refactored AdvancedSecurityMiddleware into a dynamic, Redis-backed threat detection engine with sliding-window rate tracking and signature matching.
- Sub-Accounts: Fully operationalized the sub-account transactional ledger by implementing the `request_customer_withdrawal` flow and activating multi-tier permission checks.
- Privacy: Integrated systematic PII masking (emails and wallet addresses) into notifications, audit logs, and dashboard views to ensure privacy-by-design.
- Optimization: Enabled real-time Ethereum gas monitoring via WebSockets and simplified Solana fee logic to a static high-efficiency default.
- Stability: Resolved all pending Rust compiler errors and Clippy lints across the middleware and service layers.

Files impacted: advanced_security.rs, merchant_customer_service.rs, gas_websocket_service.rs, gas_fee_service.rs, withdrawal_processor.rs, and handlers.

+++ backend/src/services/payment_service.rs
@@ -256,7 +256,7 @@ impl PaymentService {
     ) -> Result<PaymentList, PaymentServiceError> {
         let page = filters.page.unwrap_or(1).max(1);
         let page_size = filters.page_size.unwrap_or(20).clamp(1, 100);
-        let offset = (page as i64 - 1) * page_size as i64;
+        let offset = (page - 1) * page_size;
 
         // Build the base query
         let mut query = String::from("SELECT * FROM payment_transactions WHERE merchant_id = $1");
